Improved the validation process, it keeps the output files if any query is wrong

+++ queryExecuter.py
@@ -90,11 +90,14 @@ def do_the_work(cursor, repetitions, machine_type, dbms, scale_factor):
 def validate_results(machine_type, dbms, scale_factor):
     # Run the perl script
     os.system('bash %s %s %s %s > %s' % (os.path.join('answers', 'cmpall.sh'), machine_type, dbms, scale_factor, os.path.join('results', 'tmp.txt')))
-    # delete the logs, not needed
-    for i in range(23):
-        if os.path.exists('analysis_%d.log' % (i + 1)):
-            os.remove('analysis_%d.log' % (i + 1))
-    return filecmp.cmp(os.path.join('answers', 'curatedAnswers.txt'), os.path.join('results', 'tmp.txt'))
+    files_identical = filecmp.cmp(os.path.join('answers', 'curatedAnswers.txt'), os.path.join('results', 'tmp.txt'))
+
+    if not files_identical:
+        # delete the logs, not if the answers are correct
+        for i in range(23):
+            if os.path.exists('analysis_%d.log' % (i + 1)):
+                os.remove('analysis_%d.log' % (i + 1))
+    return files_identical
 
 
 if __name__ == '__main__':
@@ -112,9 +115,9 @@ if __name__ == '__main__':
     reps = 30  # preferably 30, but you can decrease this during debugging
 
     # Create connection
-    cursor = open_connection(db_username, db_password, db_hostname, database)
-
-    do_the_work(cursor, reps, machine_type, dbms, scale_factor)
+    # cursor = open_connection(db_username, db_password, db_hostname, database)
+    #
+    # do_the_work(cursor, reps, machine_type, dbms, scale_factor)
 
     title = 'Results of %s performing on %dGB data with %s' % (dbms, scale_factor, name_in_plot)
